refactor: move pretty printin to LineWriter, simplify folder structure
