+++ package.json
@@ -41,7 +41,6 @@
     ]
   },
   "dependencies": {
-    "chalk": "^2.3.0",
-    "ms": "^2.0.0"
+    "chalk": "^2.3.0"
   }
 }
